B3 and Air should simplify CFGs https://bugs.webkit.org/show_bug.cgi?id=150960
Reviewed by Geoffrey Garen.

This adds CFG simplification to both B3 and Air.

In B3, the simplification is done inside the B3::reduceStrength() fixpoint because we expect
that it will help to reveal more optimization opportunities. This is going to be particularly
true when we add Phi elimination.

In Air, the simplification is its own phase. We expect it to produce most of its benefits once
we have coalescing. Then, CFG simplification in Air will unbreak critial edges.

* JavaScriptCore.xcodeproj/project.pbxproj:
* assembler/AbortReason.h:
* assembler/MacroAssembler.h:
(JSC::MacroAssembler::oops): Reveal this as a method so that we can have an Oops instruction.
* b3/B3BasicBlock.h:
(JSC::B3::BasicBlock::predecessor):
(JSC::B3::BasicBlock::predecessors):
(JSC::B3::BasicBlock::containsPredecessor):
* b3/B3BasicBlockUtils.h: Bunch of fixes for blocks being killed.
(JSC::B3::replacePredecessor):
(JSC::B3::resetReachability):
* b3/B3ReduceStrength.cpp: Implement B3 CFG simplification.
* b3/B3ReduceStrength.h:
* b3/air/AirBasicBlock.h:
(JSC::B3::Air::BasicBlock::resize):
(JSC::B3::Air::BasicBlock::insts):
(JSC::B3::Air::BasicBlock::appendInst):
(JSC::B3::Air::BasicBlock::containsPredecessor):
* b3/air/AirGenerate.cpp:
(JSC::B3::Air::generate):
* b3/air/AirInst.cpp:
(JSC::B3::Air::Inst::hasArgEffects):
(JSC::B3::Air::Inst::dump):
* b3/air/AirInst.h:
* b3/air/AirLiveness.h:
(JSC::B3::Air::Liveness::Liveness): Fix for when blocks were killed.
* b3/air/AirOpcode.opcodes:
* b3/air/AirSimplifyCFG.cpp: Added.
(JSC::B3::Air::simplifyCFG):
* b3/air/AirSimplifyCFG.h: Added.



Canonical link: https://commits.webkit.org/169198@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@192121 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/b3/B3BasicBlockUtils.h
@@ -64,14 +64,11 @@ bool removePredecessor(BasicBlock* block, BasicBlock* predecessor)
 template<typename BasicBlock>
 bool replacePredecessor(BasicBlock* block, BasicBlock* from, BasicBlock* to)
 {
-    auto& predecessors = block->predecessors();
-    for (unsigned i = 0; i < predecessors.size(); ++i) {
-        if (predecessors[i] == from) {
-            predecessors[i] = to;
-            return true;
-        }
-    }
-    return false;
+    bool changed = false;
+    // We do it this way because 'to' may already be a predecessor of 'block'.
+    changed |= removePredecessor(block, from);
+    changed |= addPredecessor(block, to);
+    return changed;
 }
 
 // This recomputes predecessors and removes blocks that aren't reachable.
@@ -80,8 +77,10 @@ void resetReachability(
     Vector<std::unique_ptr<BasicBlock>>& blocks, const DeleteFunctor& deleteFunctor)
 {
     // Clear all predecessor lists first.
-    for (auto& block : blocks)
-        block->predecessors().resize(0);
+    for (auto& block : blocks) {
+        if (block)
+            block->predecessors().resize(0);
+    }
 
     GraphNodeWorklist<BasicBlock*, IndexSet<BasicBlock>> worklist;
     worklist.push(blocks[0].get());
@@ -93,6 +92,8 @@ void resetReachability(
     }
 
     for (unsigned i = 1; i < blocks.size(); ++i) {
+        if (!blocks[i])
+            continue;
         if (blocks[i]->predecessors().isEmpty()) {
             deleteFunctor(blocks[i].get());
             blocks[i] = nullptr;
